requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
djangorestframework==3.16.1
dotenv==0.9.9
google-auth==2.41.1
google-auth-oauthlib==1.2.3
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
sqlparse==0.5.4
stripe==14.0.1
Twisted==25.5.0
txaio==25.12.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
zope.interface==8.1.1
